Fix the logic for crates/barrels without an item
The probability of barrels not having an item was 41/101, when it should be
31/101.

Barrels and crates without an item had a 23/101 chance to be completely empty,
and a 2/101 chance to hold 30 coins (50 with ring of luck). It should instead
be 23/101 chance to hold 30 coins (50 with ring of luck), and 2/101 to hold 50
coins.

With those modifications, the expected amount of gold is 43809 per 10000
crates, and 33124 per 10000 barrels. I tested in ND and got 44070 and 32930
respectively, confirming those are the correct probabilities (or very close).